requirements.txt
aiogram==3.7.0
pillow==10.3.0
openai==1.30.0
aiohttp==3.9.5
rembg==2.0.56
python-dotenv==1.0.1
